+++ themes/article/app/routes/$slug.tsx
@@ -15,13 +15,14 @@ import {
 } from '@myst-theme/site';
 import { FrontmatterBlock } from '@myst-theme/frontmatter';
 import { getPage } from '~/utils/loaders.server';
-import { NavLink, useLoaderData, useLocation } from '@remix-run/react';
+import { useLoaderData, useLocation } from '@remix-run/react';
 import type { SiteManifest } from 'myst-config';
 import {
   ReferencesProvider,
   TabStateProvider,
   UiStateProvider,
   useLinkProvider,
+  useNavLinkProvider,
   useSiteManifest,
 } from '@myst-theme/providers';
 import type { GenericParent } from 'myst-common';
@@ -45,8 +46,8 @@ export const meta: MetaFunction = (args) => {
 export const links: LinksFunction = () => [KatexCSS];
 
 export const loader: LoaderFunction = async ({ params, request }) => {
-  const { project, slug } = params;
-  return getPage(request, { project, slug, redirect: true });
+  const { slug } = params;
+  return getPage(request, { slug, redirect: true });
 };
 
 export function ArticlePageAndNavigation({
@@ -70,14 +71,15 @@ export function ArticlePageAndNavigation({
 function ArticleNavigation() {
   const site = useSiteManifest();
   const Link = useLinkProvider();
+  const NavLink = useNavLinkProvider();
   const { pathname } = useLocation();
   const project = site?.projects?.[0];
-  const exact = pathname === `/${project?.slug}`;
+  const exact = pathname === `/`;
   return (
     <nav className="col-page-inset">
       <div className="border-y border-gray-200 py-3 mt-4 mb-6 flex flex-row justify-around">
         <Link
-          to={`/${project?.slug}`}
+          to={`/`}
           prefetch="intent"
           className={classNames('no-underline', { 'text-blue-600': exact })}
         >
@@ -86,20 +88,18 @@ function ArticleNavigation() {
         {project?.pages
           .filter((p) => 'slug' in p)
           .map((p) => {
-            if (p.level === 1)
-              return (
-                <NavLink
-                  key={p.slug}
-                  to={`/${project?.slug}/${p.slug}`}
-                  prefetch="intent"
-                  className={({ isActive }) =>
-                    classNames('no-underline', { 'text-blue-600': isActive })
-                  }
-                >
-                  {p.title}
-                </NavLink>
-              );
-            return null;
+            return (
+              <NavLink
+                key={p.slug}
+                to={`/${p.slug}`}
+                prefetch="intent"
+                className={({ isActive }) =>
+                  classNames('no-underline', { 'text-blue-600': isActive })
+                }
+              >
+                {p.title}
+              </NavLink>
+            );
           })}
       </div>
     </nav>