+++ themes/article/app/routes/_index.tsx
@@ -1,18 +1,14 @@
-import { responseNoArticle, responseNoSite } from '@myst-theme/site';
-import type { LoaderFunction } from '@remix-run/node';
-import { redirect } from '@remix-run/node';
-import { Outlet } from '@remix-run/react';
-import { getConfig } from '~/utils/loaders.server';
+import { ProjectPageCatchBoundary } from '@myst-theme/site';
+import Page, { ArticlePageAndNavigation } from './$slug';
+export { loader, links } from './$slug';
+export default Page;
 
-export const loader: LoaderFunction = async (): Promise<Response | null> => {
-  const config = await getConfig();
-  if (!config) throw responseNoSite();
-  const project = config?.projects?.[0];
-  if (!project) throw responseNoArticle();
-  return redirect(`/${project.slug}`);
-};
-
-// Note this is necessary to propagate catch boundaries, even though there is a redirect
-export default function Index() {
-  return <Outlet />;
+export function CatchBoundary() {
+  return (
+    <ArticlePageAndNavigation>
+      <main className="article-content">
+        <ProjectPageCatchBoundary />
+      </main>
+    </ArticlePageAndNavigation>
+  );
 }