+++ themes/book/app/root.tsx
@@ -9,10 +9,7 @@ import {
   getMetaTagsForSite,
   getThemeSession,
 } from '@myst-theme/site';
-export {
-  AppCatchBoundary as CatchBoundary,
-  AppDebugErrorBoundary as ErrorBoundary,
-} from '@myst-theme/site';
+export { AppCatchBoundary as CatchBoundary } from '@myst-theme/site';
 
 export const meta: MetaFunction = ({ data }) => {
   return getMetaTagsForSite({
@@ -46,6 +43,8 @@ export const loader: LoaderFunction = async ({ request }): Promise<SiteLoader> =
     theme: themeSession.getTheme(),
     config,
     CONTENT_CDN_PORT: process.env.CONTENT_CDN_PORT ?? 3100,
+    MODE: (process.env.MODE ?? 'app') as 'app' | 'static',
+    BASE_URL: process.env.BASE_URL || undefined,
   };
   return data;
 };